+++ src/EngineManager.php
@@ -8,6 +8,7 @@ use Algolia\AlgoliaSearch\Support\UserAgent;
 use Exception;
 use Illuminate\Support\Manager;
 use Laravel\Scout\Engines\AlgoliaEngine;
+use Laravel\Scout\Engines\CollectionEngine;
 use Laravel\Scout\Engines\MeiliSearchEngine;
 use Laravel\Scout\Engines\NullEngine;
 use MeiliSearch\Client as MeiliSearch;
@@ -34,7 +35,7 @@ class EngineManager extends Manager
     {
         $this->ensureAlgoliaClientIsInstalled();
 
-        UserAgent::addCustomUserAgent('Laravel Scout', '9.0.0-dev');
+        UserAgent::addCustomUserAgent('Laravel Scout', '10.0.0-dev');
 
         $config = SearchConfig::create(
             config('scout.algolia.id'),
@@ -42,6 +43,18 @@ class EngineManager extends Manager
         )->setDefaultHeaders(
             $this->defaultAlgoliaHeaders()
         );
+
+        if (is_int($connectTimeout = config('scout.algolia.connect_timeout'))) {
+            $config->setConnectTimeout($connectTimeout);
+        }
+
+        if (is_int($readTimeout = config('scout.algolia.read_timeout'))) {
+            $config->setReadTimeout($readTimeout);
+        }
+
+        if (is_int($writeTimeout = config('scout.algolia.write_timeout'))) {
+            $config->setWriteTimeout($writeTimeout);
+        }
 
         return new AlgoliaEngine(Algolia::createWithConfig($config), config('scout.soft_delete'));
     }
@@ -124,7 +137,17 @@ class EngineManager extends Manager
     }
 
     /**
-     * Create a Null engine instance.
+     * Create a collection engine instance.
+     *
+     * @return \Laravel\Scout\Engines\CollectionEngine
+     */
+    public function createCollectionDriver()
+    {
+        return new CollectionEngine;
+    }
+
+    /**
+     * Create a null engine instance.
      *
      * @return \Laravel\Scout\Engines\NullEngine
      */