+++ tests/Unit/AlgoliaEngineTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Laravel\Scout\Tests;
+namespace Laravel\Scout\Tests\Unit;
 
 use Algolia\AlgoliaSearch\SearchClient;
 use Illuminate\Database\Eloquent\Collection;
@@ -20,6 +20,7 @@ class AlgoliaEngineTest extends TestCase
     protected function setUp(): void
     {
         Config::shouldReceive('get')->with('scout.after_commit', m::any())->andReturn(false);
+        Config::shouldReceive('get')->with('scout.soft_delete', m::any())->andReturn(false);
     }
 
     protected function tearDown(): void
@@ -64,15 +65,29 @@ class AlgoliaEngineTest extends TestCase
         $engine->search($builder);
     }
 
+    public function test_search_sends_correct_parameters_to_algolia_for_where_in_search()
+    {
+        $client = m::mock(SearchClient::class);
+        $client->shouldReceive('initIndex')->with('table')->andReturn($index = m::mock(stdClass::class));
+        $index->shouldReceive('search')->with('zonda', [
+            'numericFilters' => ['foo=1', ['bar=1', 'bar=2']],
+        ]);
+
+        $engine = new AlgoliaEngine($client);
+        $builder = new Builder(new SearchableModel, 'zonda');
+        $builder->where('foo', 1)->whereIn('bar', [1, 2]);
+        $engine->search($builder);
+    }
+
     public function test_map_correctly_maps_results_to_models()
     {
         $client = m::mock(SearchClient::class);
         $engine = new AlgoliaEngine($client);
 
         $model = m::mock(stdClass::class);
-        $model->shouldReceive('queryScoutModelsByIds->cursor')->andReturn(new LazyCollection($models = Collection::make([
+        $model->shouldReceive('getScoutModelsByIds')->andReturn($models = Collection::make([
             new SearchableModel(['id' => 1]),
-        ])));
+        ]));
 
         $builder = m::mock(Builder::class);
 
@@ -89,16 +104,69 @@ class AlgoliaEngineTest extends TestCase
         $engine = new AlgoliaEngine($client);
 
         $model = m::mock(stdClass::class);
-        $model->shouldReceive('queryScoutModelsByIds->cursor')->andReturn(new LazyCollection($models = Collection::make([
+        $model->shouldReceive('getScoutModelsByIds')->andReturn($models = Collection::make([
             new SearchableModel(['id' => 1]),
             new SearchableModel(['id' => 2]),
             new SearchableModel(['id' => 3]),
             new SearchableModel(['id' => 4]),
-        ])));
+        ]));
 
         $builder = m::mock(Builder::class);
 
         $results = $engine->map($builder, ['nbHits' => 4, 'hits' => [
+            ['objectID' => 1, 'id' => 1],
+            ['objectID' => 2, 'id' => 2],
+            ['objectID' => 4, 'id' => 4],
+            ['objectID' => 3, 'id' => 3],
+        ]], $model);
+
+        $this->assertCount(4, $results);
+
+        // It's important we assert with array keys to ensure
+        // they have been reset after sorting.
+        $this->assertEquals([
+            0 => ['id' => 1],
+            1 => ['id' => 2],
+            2 => ['id' => 4],
+            3 => ['id' => 3],
+        ], $results->toArray());
+    }
+
+    public function test_lazy_map_correctly_maps_results_to_models()
+    {
+        $client = m::mock(SearchClient::class);
+        $engine = new AlgoliaEngine($client);
+
+        $model = m::mock(stdClass::class);
+        $model->shouldReceive('queryScoutModelsByIds->cursor')->andReturn($models = LazyCollection::make([
+            new SearchableModel(['id' => 1]),
+        ]));
+
+        $builder = m::mock(Builder::class);
+
+        $results = $engine->lazyMap($builder, ['nbHits' => 1, 'hits' => [
+            ['objectID' => 1, 'id' => 1],
+        ]], $model);
+
+        $this->assertCount(1, $results);
+    }
+
+    public function test_lazy_map_method_respects_order()
+    {
+        $client = m::mock(SearchClient::class);
+        $engine = new AlgoliaEngine($client);
+
+        $model = m::mock(stdClass::class);
+        $model->shouldReceive('queryScoutModelsByIds->cursor')->andReturn($models = LazyCollection::make([
+            new SearchableModel(['id' => 1]),
+            new SearchableModel(['id' => 2]),
+            new SearchableModel(['id' => 3]),
+            new SearchableModel(['id' => 4]),
+        ]));
+
+        $builder = m::mock(Builder::class);
+
+        $results = $engine->lazyMap($builder, ['nbHits' => 4, 'hits' => [
             ['objectID' => 1, 'id' => 1],
             ['objectID' => 2, 'id' => 2],
             ['objectID' => 4, 'id' => 4],